requirements.txt
python-lorem>=1.1.2
psycopg2>=2.8.5
